libpng: Upgrade to 1.6.25
The changes from 1.6.22 (beta) to
1.6.25 (final) include many general fixes,
fixes for memory leak issues and optimizations
in order to improve performance.
Also upstream added support for the MIPS
platform.

Support for Intel SSE isn't built in any more.
For support a patch in contrib/intel/intel_sse.patch
has to be applied. This patch additionally
documents this in README.android as an Android
specific addition.

The patch for Intel SSE support
from upstream applied cleanly.

Libpng 1.6.25 released on September 1, 2016.

Test: built for ARM and checked whether NEON optimizations were compiled
      built for x86/x86_64 and checked whether SSE optimizations were compiled

Change-Id: If647f42a1fb4d7022286010d262d287dac4030df
Signed-off-by: Alex Naidis <alex.naidis@linux.com>